refact(input/) Refact input component
- Refact the Input.component.js
- Move style to different Input.component.style.js
